requirements.txt
flask
redis
kafka-python
pandas